Refactor WOPR2 into multiple files
Refactor `WOPR2.py` into multiple files for better maintainability and navigation.

* **Added `utility.py`**
  - Contains `Utility` class with `remove_duplicates` and `reverse_lookup_submarine` functions.

* **Added `draw.py`**
  - Contains `Draw` class with methods for drawing characters, targets, fallout, circles, lines, and handling console output.

* **Added `missiles.py`**
  - Contains `Missiles` class with methods for calculating distances and launching missiles using different algorithms.

* **Added `map.py`**
  - Contains `Map` class with methods for printing the map, handling ocean tiles, checking for obstructions, converting coordinates, and printing submarines and silos.

* **Modified `game.py`**
  - Imports necessary classes and functions from the new files.
  - Contains `classic_mode`, `SubsAndSilos`, `submarinesandsilos`, and `main` functions for game logic and handling player and enemy moves.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/illegalbyte/WOPR?shareId=XXXX-XXXX-XXXX-XXXX).
